+++ hedera-node/hedera-app/src/main/java/com/hedera/node/app/records/impl/producers/formats/BlockRecordWriterFactoryImpl.java
@@ -8,12 +8,12 @@ import com.hedera.node.app.records.impl.producers.BlockRecordWriterFactory;
 import com.hedera.node.app.records.impl.producers.formats.v6.BlockRecordWriterV6;
 import com.hedera.node.config.ConfigProvider;
 import com.hedera.node.config.data.BlockRecordStreamConfig;
-import com.swirlds.common.stream.Signer;
 import com.swirlds.state.lifecycle.info.NodeInfo;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.nio.file.FileSystem;
 import javax.inject.Inject;
 import javax.inject.Singleton;
+import org.hiero.base.crypto.Signer;
 
 @Singleton
 public class BlockRecordWriterFactoryImpl implements BlockRecordWriterFactory {
@@ -49,11 +49,12 @@ public class BlockRecordWriterFactoryImpl implements BlockRecordWriterFactory {
 
         // pick a record file format
         return switch (recordFileVersion) {
-            case 6 -> new BlockRecordWriterV6(
-                    configProvider.getConfiguration().getConfigData(BlockRecordStreamConfig.class),
-                    selfNodeInfo,
-                    signer,
-                    fileSystem);
+            case 6 ->
+                new BlockRecordWriterV6(
+                        configProvider.getConfiguration().getConfigData(BlockRecordStreamConfig.class),
+                        selfNodeInfo,
+                        signer,
+                        fileSystem);
             case 7 -> throw new IllegalArgumentException("Record file version 7 is not yet supported");
             default -> throw new IllegalArgumentException("Unknown record file version: " + recordFileVersion);
         };
