+++ platform-sdk/consensus-utility/src/main/java/module-info.java
@@ -1,6 +1,8 @@
 // SPDX-License-Identifier: Apache-2.0
 module org.hiero.consensus.utility {
     exports org.hiero.consensus.config;
+    exports org.hiero.consensus.crypto;
+    exports org.hiero.consensus.exceptions;
     exports org.hiero.consensus.roster;
 
     requires transitive com.hedera.node.hapi;
@@ -10,5 +12,7 @@ module org.hiero.consensus.utility {
     requires transitive org.hiero.base.crypto;
     requires transitive org.hiero.consensus.model;
     requires com.swirlds.base;
+    requires com.swirlds.logging;
+    requires org.bouncycastle.provider;
     requires static transitive com.github.spotbugs.annotations;
 }
