+++ platform-sdk/platform-apps/tests/StressTestingTool/build.gradle.kts
@@ -7,6 +7,7 @@ mainModuleInfo { annotationProcessor("com.swirlds.config.processor") }
 
 testModuleInfo {
     requires("org.hiero.base.crypto")
+    requires("org.hiero.consensus.utility")
     requires("org.assertj.core")
     requires("org.junit.jupiter.api")
     requires("org.junit.jupiter.params")
