+++ platform-sdk/platform-apps/tests/StressTestingTool/src/test/java/com/swirlds/demo/stress/StressTestingToolStateTest.java
@@ -18,7 +18,6 @@ import com.swirlds.common.test.fixtures.Randotron;
 import com.swirlds.config.api.ConfigurationBuilder;
 import com.swirlds.platform.crypto.KeyGeneratingException;
 import com.swirlds.platform.crypto.KeysAndCertsGenerator;
-import com.swirlds.platform.crypto.PlatformSigner;
 import com.swirlds.platform.crypto.PublicStores;
 import com.swirlds.platform.state.PlatformStateModifier;
 import com.swirlds.platform.system.InitTrigger;
@@ -32,6 +31,7 @@ import java.util.Collections;
 import java.util.List;
 import java.util.function.Consumer;
 import org.hiero.base.crypto.Hash;
+import org.hiero.consensus.crypto.PlatformSigner;
 import org.hiero.consensus.model.event.ConsensusEvent;
 import org.hiero.consensus.model.event.PlatformEvent;
 import org.hiero.consensus.model.hashgraph.Round;
