+++ platform-sdk/swirlds-common/src/testFixtures/java/com/swirlds/common/test/fixtures/stream/StreamFileSigner.java
@@ -5,7 +5,6 @@ import static com.swirlds.logging.legacy.LogMarker.EXCEPTION;
 import static org.hiero.base.crypto.internal.DetRandomProvider.getDetRandom;
 import static org.hiero.base.utility.CommonUtils.hex;
 
-import com.swirlds.common.stream.Signer;
 import com.swirlds.common.utility.CommonUtils;
 import java.security.InvalidKeyException;
 import java.security.KeyPair;
@@ -21,6 +20,7 @@ import org.apache.logging.log4j.Logger;
 import org.apache.logging.log4j.Marker;
 import org.apache.logging.log4j.MarkerManager;
 import org.hiero.base.crypto.SignatureType;
+import org.hiero.base.crypto.Signer;
 
 public class StreamFileSigner implements Signer {
     public static final String SIG_ALGORITHM = SignatureType.RSA.signingAlgorithm();
