+++ platform-sdk/swirlds-common/src/testFixtures/java/com/swirlds/common/test/fixtures/stream/StreamObjectWorker.java
@@ -8,7 +8,6 @@ import com.swirlds.common.stream.HashCalculatorForStream;
 import com.swirlds.common.stream.QueueThreadObjectStream;
 import com.swirlds.common.stream.QueueThreadObjectStreamConfiguration;
 import com.swirlds.common.stream.RunningHashCalculatorForStream;
-import com.swirlds.common.stream.Signer;
 import com.swirlds.common.stream.internal.TimestampStreamFileWriter;
 import java.io.IOException;
 import java.security.NoSuchAlgorithmException;
@@ -17,6 +16,7 @@ import java.util.Deque;
 import java.util.Iterator;
 import java.util.LinkedList;
 import org.hiero.base.crypto.Hash;
+import org.hiero.base.crypto.Signer;
 import org.hiero.base.io.SelfSerializable;
 import org.hiero.base.io.streams.SerializableDataOutputStream;
