+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/crypto/PublicStores.java
@@ -11,6 +11,7 @@ import java.security.cert.X509Certificate;
 import java.util.Objects;
 import java.util.stream.Stream;
 import org.hiero.base.crypto.CryptographyException;
+import org.hiero.consensus.crypto.CryptoConstants;
 import org.hiero.consensus.model.node.NodeId;
 
 /**
@@ -91,7 +92,8 @@ public record PublicStores(KeyStore sigTrustStore, KeyStore agrTrustStore) {
         try {
             certificate = switch (type) {
                 case SIGNING -> sigTrustStore.getCertificate(name);
-                case AGREEMENT -> agrTrustStore.getCertificate(name);};
+                case AGREEMENT -> agrTrustStore.getCertificate(name);
+            };
         } catch (KeyStoreException e) {
             // cannot be thrown because we ensure the key store is initialized in the constructor
             throw new CryptographyException(e);
