+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/event/creation/tipset/TipsetEventCreator.java
@@ -8,7 +8,6 @@ import com.hedera.hapi.node.base.SemanticVersion;
 import com.hedera.hapi.node.state.roster.Roster;
 import com.swirlds.base.time.Time;
 import com.swirlds.common.context.PlatformContext;
-import com.swirlds.common.stream.HashSigner;
 import com.swirlds.common.utility.throttle.RateLimitedLogger;
 import com.swirlds.platform.event.EventUtils;
 import com.swirlds.platform.event.hashing.PbjStreamHasher;
@@ -25,6 +24,7 @@ import java.util.Objects;
 import java.util.Random;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.hiero.base.crypto.Hash;
 import org.hiero.base.crypto.Signature;
 import org.hiero.consensus.config.EventConfig;
 import org.hiero.consensus.event.creator.impl.EventCreator;
@@ -456,5 +456,18 @@ public class TipsetEventCreator implements EventCreator {
         }
 
         return sb.toString();
+    }
+
+    /**
+     * Capable of signing a {@link Hash}
+     */
+    @FunctionalInterface
+    public interface HashSigner {
+        /**
+         * @param hash
+         * 		the hash to sign
+         * @return the signature for the hash provided
+         */
+        Signature sign(Hash hash);
     }
 }
