+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/event/signing/DefaultSelfEventSigner.java
@@ -1,10 +1,10 @@
 // SPDX-License-Identifier: Apache-2.0
 package com.swirlds.platform.event.signing;
 
-import com.swirlds.platform.crypto.PlatformSigner;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.util.Objects;
 import org.hiero.base.crypto.Signature;
+import org.hiero.consensus.crypto.PlatformSigner;
 import org.hiero.consensus.model.event.PlatformEvent;
 import org.hiero.consensus.model.event.UnsignedEvent;
 import org.hiero.consensus.model.node.KeysAndCerts;
