+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/state/signer/DefaultStateSigner.java
@@ -3,12 +3,12 @@ package com.swirlds.platform.state.signer;
 
 import com.hedera.hapi.platform.event.StateSignatureTransaction;
 import com.hedera.pbj.runtime.io.buffer.Bytes;
-import com.swirlds.platform.crypto.PlatformSigner;
 import com.swirlds.platform.state.signed.ReservedSignedState;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import edu.umd.cs.findbugs.annotations.Nullable;
 import java.util.Objects;
 import org.hiero.base.crypto.Hash;
+import org.hiero.consensus.crypto.PlatformSigner;
 
 /**
  * A standard implementation of a {@link StateSigner}.
