+++ platform-sdk/swirlds-platform-core/src/test/java/com/swirlds/platform/crypto/KeysAndCertsTest.java
@@ -14,6 +14,7 @@ import java.util.Random;
 import org.hiero.base.crypto.KeyType;
 import org.hiero.base.crypto.Signature;
 import org.hiero.base.crypto.test.fixtures.PreGeneratedPublicKeys;
+import org.hiero.consensus.crypto.PlatformSigner;
 import org.hiero.consensus.model.node.KeysAndCerts;
 import org.hiero.consensus.model.node.NodeId;
 import org.hiero.consensus.roster.RosterUtils;
