+++ platform-sdk/swirlds-platform-core/src/test/java/com/swirlds/platform/event/tipset/TipsetEventCreatorTestUtils.java
@@ -17,10 +17,10 @@ import com.hedera.hapi.node.state.roster.RosterEntry;
 import com.hedera.pbj.runtime.io.buffer.Bytes;
 import com.swirlds.base.time.Time;
 import com.swirlds.common.context.PlatformContext;
-import com.swirlds.common.stream.HashSigner;
 import com.swirlds.common.test.fixtures.platform.TestPlatformContextBuilder;
 import com.swirlds.platform.event.creation.tipset.ChildlessEventTracker;
 import com.swirlds.platform.event.creation.tipset.TipsetEventCreator;
+import com.swirlds.platform.event.creation.tipset.TipsetEventCreator.HashSigner;
 import com.swirlds.platform.event.creation.tipset.TipsetTracker;
 import com.swirlds.platform.event.creation.tipset.TipsetWeightCalculator;
 import com.swirlds.platform.event.orphan.DefaultOrphanBuffer;
