+++ platform-sdk/swirlds-platform-core/src/test/java/com/swirlds/platform/state/signer/StateSignerTests.java
@@ -12,12 +12,12 @@ import static org.mockito.Mockito.when;
 import com.hedera.hapi.platform.event.StateSignatureTransaction;
 import com.swirlds.common.test.fixtures.Randotron;
 import com.swirlds.merkledb.MerkleDb;
-import com.swirlds.platform.crypto.PlatformSigner;
 import com.swirlds.platform.state.signed.ReservedSignedState;
 import com.swirlds.platform.state.signed.SignedState;
 import com.swirlds.platform.test.fixtures.state.RandomSignedStateGenerator;
 import org.hiero.base.crypto.Hash;
 import org.hiero.base.crypto.Signature;
+import org.hiero.consensus.crypto.PlatformSigner;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
